third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbcursor_continue_objectstore4-expected.txt

This is a testharness.js-based test.
PASS IDBCursor.continue() - object store - attempt to iterate to the next record when the direction is set for the previous record  
Harness: the test ran to completion.
